backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.10",
    "setuptools>=80.9.0",
    "celery>=5.5.3",
    "redis>=6.2.0",
]
